+++ doc/references/release-notes.rst
@@ -48,6 +48,8 @@ Features
 
 * Add additional standard line types from pandapower.
 
+* New network indexing methods: ``n.get_network``, ``n.get_scenario``, ``n.slice_network`` and enhanced ``n[...]`` for convenient network access.
+
 
 * The ``Network.add()`` method now returns ``None`` by default. Use ``return_names=True`` 
   to get the previous behavior of returning component names which have been added.